requirements.txt
albumentations==1.3.1
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
einops==0.3.0
fairscale==0.4.4
faiss-gpu==1.7.2
imgaug==0.4.0
importlib-metadata==4.11.3
iopath
ipdb
jupyter==1.0.0
kaggle==1.5.13
lightning-utilities==0.9.0
opencv-python
opencv-python-headless
pandas
pycocoevalcap==1.2
pycocotools==2.0.6
pytorch-lightning==1.9.1
pytorchyolo==1.6.2
PyYAML==6.0
scikit-image==0.18.1
scikit-learn==1.0.2
scipy
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
timm
torchmetrics==0.11.4
torchsummary==1.5.1
tqdm==4.65.0
transformers==4.30.2
wandb==0.15.8
warmup-scheduler==0.3
nltk
datasets
lightning
segmentation-models-pytorch
sahi
imantics
pydicom
mdai
pytorch_pretrained_vit